testing/unit_test_aws/requirements.txt
adal==1.2.2
boto3==1.26.6
botocore==1.29.6
certifi==2022.12.7
cffi==1.16.0
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.9
jmespath==0.10.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.28.1
s3transfer==0.6.0
six==1.14.0
urllib3==1.26.5
xmlrunner==1.7.7
